requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
unstructured
bs4
arxiv
pymupdf
wikipedia
python-dotenv
langchain_openai
chromadb
langchain_text_splitters
faiss-cpu
langchain_core
langchain_groq
langserve
fastapi
sse_starlette
uvicorn